Cleanup of methods and defaults.
